docker/pyproject.toml
[project]
name = "nobber-emote-fetch"
version = "0.1.0"
description = "Automation to get number of certain emotes and store them in a JSON file"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-twitch-client>=0.7.1",
    "tzdata>=2025.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.12.9",
]
